teste çıkan kod

+++ Lyra/Telemetry - Electric/yeniform/Variables/DataGPS.cs
@@ -8,8 +8,8 @@ namespace Telemetri.Variables
 {
     public static class DataGPS
     {
-        public static float latitude_f32;      // uint64 / 1000000
-        public static float longtitude_f32;    // uint64 / 1000000
+        public static double latitude_f32;      // uint64 / 1000000
+        public static double longtitude_f32;    // uint64 / 1000000
         public static byte speed_u8;
         public static byte sattelite_u8;
         public static byte efficiency_u8;
